+++ lib/quack_db/cursor.ex
@@ -8,3 +8,17 @@ defmodule QuackDB.Cursor do
 
   defstruct [:ref, :result_uuid, :columns, :result_types, :connection_id]
 end
+
+defimpl Inspect, for: QuackDB.Cursor do
+  import Inspect.Algebra
+
+  def inspect(cursor, opts) do
+    fields = [
+      result_uuid: cursor.result_uuid,
+      columns: cursor.columns,
+      connection_id: QuackDB.Inspect.short_id(cursor.connection_id)
+    ]
+
+    concat(QuackDB.Inspect.container("QuackDB.Cursor", fields, opts))
+  end
+end